+++ server/controllers/memoController.js
@@ -1,11 +1,20 @@
 const memoService = require("../services/memoService.js");
+const errorHandler = require("../middlewares/errorHandler.js");
 
+const commonErrors = require("../middlewares/commonError.js");
 const memoController = {
 	//post
-	async postMemo(req, res) {
-		const dogId = req.params.id;
-		const memoData = req.body;
+	async postMemo(req, res, next) {
 		try {
+			const dogId = req.params.id;
+			const memoData = req.body;
+			if (!dogId || !memoData) {
+				throw new errorHandler(
+					commonErrors.argumentError,
+					"데이터를 받아오지 못했습니다.",
+					{ statusCode: 400 }
+				);
+			}
 			await memoService.createMemo(dogId, memoData, req.currentUserId);
 			res.status(201).json({ message: "Data created successfully" });
 		} catch (error) {
@@ -13,9 +22,16 @@ const memoController = {
 		}
 	},
 	// get
-	async getMemoById(req, res) {
-		const dogId = req.params.id;
+	async getMemoById(req, res, next) {
 		try {
+			const dogId = req.params.id;
+			if (!dogId) {
+				throw new errorHandler(
+					commonErrors.argumentError,
+					"데이터를 받아오지 못했습니다.",
+					{ statusCode: 400 }
+				);
+			}
 			const memo = await memoService.getMemoById(dogId, req.currentUserId);
 			res.json(memo);
 		} catch (error) {
@@ -23,11 +39,18 @@ const memoController = {
 		}
 	},
 	//put
-	async putMemo(req, res) {
-		const dogId = req.params.id;
-		const memoId = req.params.memoId;
-		const updatedMemoData = req.body;
+	async putMemo(req, res, next) {
 		try {
+			const dogId = req.params.id;
+			const memoId = req.params.memoId;
+			const updatedMemoData = req.body;
+			if (!dogId || !memoId || !updatedMemoData) {
+				throw new errorHandler(
+					commonErrors.argumentError,
+					"데이터를 받아오지 못했습니다.",
+					{ statusCode: 400 }
+				);
+			}
 			await memoService.updateMemo(
 				dogId,
 				memoId,
@@ -41,17 +64,22 @@ const memoController = {
 	},
 
 	//delete
-	async deleteMemo(req, res) {
-		const dogId = req.params.id;
-		const memoId = req.params.memoId;
-
+	async deleteMemo(req, res, next) {
 		try {
+			const dogId = req.params.id;
+			const memoId = req.params.memoId;
+			if (!dogId || !memoId) {
+				throw new errorHandler(
+					commonErrors.argumentError,
+					"데이터를 받아오지 못했습니다.",
+					{ statusCode: 400 }
+				);
+			}
 			await memoService.deleteMemo(dogId, memoId, req.currentUserId);
-			res.status(204).json({ message: "Data deleted successfully" });
+			res.status(200).json({ message: "Data deleted successfully" });
 		} catch (error) {
 			next(error);
 		}
 	},
 };
-
 module.exports = memoController;
