+++ server/controllers/weightController.js
@@ -1,10 +1,18 @@
 const weightService = require("../services/weightService.js");
-
+const errorHandler = require("../middlewares/errorHandler.js");
+const commonErrors = require("../middlewares/commonError.js");
 const weightController = {
-	async postWeight(req, res) {
-		const dogId = req.params.id;
-		const weightData = req.body;
+	async postWeight(req, res, next) {
 		try {
+			const dogId = req.params.id;
+			const weightData = req.body;
+			if (!dogId || !weightData) {
+				throw new errorHandler(
+					commonErrors.argumentError,
+					"데이터를 받아오지 못했습니다.",
+					{ statusCode: 400 }
+				);
+			}
 			await weightService.createWeight(dogId, weightData, req.currentUserId);
 			res.status(201).json({ message: "Data created successfully" });
 		} catch (error) {
@@ -12,9 +20,16 @@ const weightController = {
 		}
 	},
 	// 몸무게데이터 가져오기
-	async getWeightById(req, res) {
-		const dogId = req.params.id;
+	async getWeightById(req, res, next) {
 		try {
+			const dogId = req.params.id;
+			if (!dogId) {
+				throw new errorHandler(
+					commonErrors.argumentError,
+					"데이터를 받아오지 못했습니다.",
+					{ statusCode: 400 }
+				);
+			}
 			const weight = await weightService.getWeightById(
 				dogId,
 				req.currentUserId
@@ -25,11 +40,18 @@ const weightController = {
 		}
 	},
 	// 몸무게 정보 업데이트
-	async putWeight(req, res) {
-		const dogId = req.params.id;
-		const weightId = req.params.weightId;
-		const updatedWeightData = req.body;
+	async putWeight(req, res, next) {
 		try {
+			const dogId = req.params.id;
+			const weightId = req.params.weightId;
+			const updatedWeightData = req.body;
+			if (!dogId || !weightId || !updatedWeightData) {
+				throw new errorHandler(
+					commonErrors.argumentError,
+					"데이터를 받아오지 못했습니다.",
+					{ statusCode: 400 }
+				);
+			}
 			await weightService.updateWeight(
 				dogId,
 				weightId,
@@ -43,11 +65,17 @@ const weightController = {
 	},
 
 	// 몸무게 삭제
-	async deleteWeight(req, res) {
-		const dogId = req.params.id;
-		const weightId = req.params.weightId;
-
+	async deleteWeight(req, res, next) {
 		try {
+			const dogId = req.params.id;
+			const weightId = req.params.weightId;
+			if (!dogId || !weightId) {
+				throw new errorHandler(
+					commonErrors.argumentError,
+					"데이터를 받아오지 못했습니다.",
+					{ statusCode: 400 }
+				);
+			}
 			await weightService.deleteWeight(dogId, weightId, req.currentUserId);
 			res.status(204).json({ message: "Data deleted successfully" });
 		} catch (error) {
